requirements.txt
numpy
scipy
scikit-learn
pyyaml

# MASLab core runtime
tqdm
openai
tenacity

# datasets builder
datasets

# SelfOrg embeddings (preferred backend)
sentence-transformers
transformers

